requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20
pydantic==2.11.10
cryptography==46.0.2
ecdsa==0.19.1
rsa==4.9.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.0
email-validator==2.1.0
PyJWT==2.8.0
